Выполнено задание: 5.16. Модуляция
